tests: Update presubmit.sh to compile ALL/gem5.fast
This part of the Kokoro presubmit tests was designed to ensure gem5
still compiled sucessfully with Clang and to the '.fast' variant. ARM
was chosen arbitarily. Now that ALL exists, it makes more sense to use
it for this test.

Change-Id: Ia3593f7dd506205da13802a69094f4dd7019ab90
Reviewed-on: https://gem5-review.googlesource.com/c/public/gem5/+/63371
Reviewed-by: Jason Lowe-Power <power.jg@gmail.com>
Maintainer: Jason Lowe-Power <power.jg@gmail.com>
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-on: https://soc-sim-external-review.googlesource.com/c/gem5/gem5/+/11322
Reviewed-by: Roger Chang <rogerycchang@google.com>
Presubmit-Ready: Roger Chang <rogerycchang@google.com>